auth-service/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
python-dotenv
PyJWT
Werkzeug==2.2.2